Attempt to fix build

+++ azure-pipelines/end-to-end-tests-dir/bundles.ps1
@@ -1,6 +1,8 @@
 . $PSScriptRoot/../end-to-end-tests-prelude.ps1
 
-Remove-Item env:VCPKG_DOWNLOADS
+if (Test-Path env:VCPKG_DOWNLOADS) {
+    Remove-Item env:VCPKG_DOWNLOADS
+}
 
 if ($IsWindows) {
     $cache_home = $env:LOCALAPPDATA
